// reolio/src/generate/schema/http/models/WordDetail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * folio
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
import type { WordDefinition } from "./WordDefinition";
import {
  WordDefinitionFromJSON,
  WordDefinitionFromJSONTyped,
  WordDefinitionToJSON,
} from "./WordDefinition";

/**
 *
 * @export
 * @interface WordDetail
 */
export interface WordDetail {
  /**
   *
   * @type {string}
   * @memberof WordDetail
   */
  word: string;
  /**
   *
   * @type {Array<WordDefinition>}
   * @memberof WordDetail
   */
  definitions: Array<WordDefinition>;
  /**
   *
   * @type {number}
   * @memberof WordDetail
   */
  frequency: number;
}

/**
 * Check if a given object implements the WordDetail interface.
 */
export function instanceOfWordDetail(value: object): boolean {
  if (!("word" in value)) return false;
  if (!("definitions" in value)) return false;
  if (!("frequency" in value)) return false;
  return true;
}

export function WordDetailFromJSON(json: any): WordDetail {
  return WordDetailFromJSONTyped(json, false);
}

export function WordDetailFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): WordDetail {
  if (json == null) {
    return json;
  }
  return {
    word: json["word"],
    definitions: (json["definitions"] as Array<any>).map(
      WordDefinitionFromJSON,
    ),
    frequency: json["frequency"],
  };
}

export function WordDetailToJSON(value?: WordDetail | null): any {
  if (value == null) {
    return value;
  }
  return {
    word: value["word"],
    definitions: (value["definitions"] as Array<any>).map(WordDefinitionToJSON),
    frequency: value["frequency"],
  };
}